abstracted the auth form to unique container
